+++ src/main/java/com/bi/oranj/model/bi/ClientAUM.java
@@ -12,5 +12,5 @@ import java.math.BigInteger;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ClientAUM extends AUMSummary {
 
-    private BigInteger clientId;
+    private Long clientId;
 }
